[wptrunner] Replace `content_shell` with `headless_shell`
Chromium CI plans to replace `content_shell` [0] with `headless_shell`
[1] for running most WPTs. Replace the product accordingly, which is
mostly a `chrome` clone.

[0]: https://chromium.googlesource.com/chromium/src/+/HEAD/docs/testing/web_tests_in_content_shell.md
[1]: https://chromium.googlesource.com/chromium/src/+/HEAD/headless/README.md
