+++ tools/wptrunner/wptrunner/browsers/__init__.py
@@ -27,10 +27,10 @@ product_list = ["android_webview",
                 "chrome_android",
                 "chrome_ios",
                 "chromium",
-                "content_shell",
                 "edge",
                 "firefox",
                 "firefox_android",
+                "headless_shell",
                 "safari",
                 "sauce",
                 "servo",
